% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BlissOwen}
\alias{BlissOwen}
\title{BlissOwen}
\usage{
BlissOwen(y)
}
\arguments{
\item{y}{A vector of observations, size p}
}
\value{
VPhi["Phi","valid"] A list with a real value, the Bliss-Owen estimation for the overdispersion parameter, and a 
boolean mark to indicate if the estimation is considered valid. It will not be when statistics u, v or both are negative or null.
}
\description{
Gets a vector of p observations assuming they come from a negative binomial distribution with overdispersion
parameter Phi (var=mu+Phi*mu^2) and uses the Bliss-Owen method to get an estimator for Phi
}
\examples{
n = 200
mu = 30
size = .7
y = rnbinom(n = n, size = size, mu= mu)
phi_est = BlissOwen(y)
}
