% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{coef.fitgen_glm}
\alias{coef.fitgen_glm}
\title{Accesor for the coefficients}
\usage{
\method{coef}{fitgen_glm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_glm}}

\item{...}{Additional arguments}
}
\description{
Accesor for the coefficients
}
