% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{coef.fitgen_Lglm}
\alias{coef.fitgen_Lglm}
\title{Method to extract the coefficients of a list
of glm's 
 Accesor for coefficients}
\usage{
\method{coef}{fitgen_Lglm}(object, type = c("full", "common"), ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{The coefficients corresponds to the full model (with
neighbours)
or to the common model (only with the common predictors)}

\item{...}{Additional arguments}
}
\description{
Method to extract the coefficients of a list
of glm's 
 Accesor for coefficients
}
