% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{coefficients.fitgen_Lglm}
\alias{coefficients.fitgen_Lglm}
\title{Accesor for coefficients}
\usage{
\method{coefficients}{fitgen_Lglm}(object, type = c("full", "common"), ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{The coefficients corresponds to the full model (with neighbours)
or to the common model (only with the common predictors)}

\item{...}{Additional arguments}
}
\description{
Accesor for coefficients
}
