% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CommonBlissOwen}
\alias{CommonBlissOwen}
\title{CommonBlissOwen}
\usage{
CommonBlissOwen(y, maxit = 100L, tol = 1e-09, posthres = 0)
}
\arguments{
\item{y}{A matrix of sample, size n x p}

\item{maxit}{Maximum number of allowed iterations. Default: 100}

\item{tol}{Tolerance to exit if change is below its value. Default: 1e-9}

\item{posthres}{Threhold to consider a sample valid if its v-value is strictly above it. Default: 0.0}
}
\value{
VPhi["Phi","valid"] A list of two element, a real number with the Phi estimated value and a  boolean array with length=nrows(y).
}
\description{
Gets a matrix of n samples of p observations each given as a nxp matrix (one sample per row) and, assuming that all 
of them come from a negative binomial ditribution with a common overdispersion Phi value (var=mu+Phi*mu^2), uses
the Bliss-Owen method to get a common estimator for Phi using all the samples for which estimation is possible.
These samples are marked as true in an array of boolean values which is returned. Estimation is considered impossible
if the v-statistic (s^2-mu) is lower or equal than a threshold (0 by default).
}
\details{
When estimation is possible, the method proceeds iteratively until either the difference of estimated values between
successive iterations is below the tolerance or until the fixed number of iterations has been reached.
}
\examples{
n = 100
N = 400
mu = 1:N
size = .7
y = t(sapply(mu,function(i) rnbinom(n = n, size = size, mu= i)))
VPhi=CommonBlissOwen(y)
cat("Estimated value: ",VPhi$Phi," using ",sum(VPhi$valid)," of all the samples.\n")
}
