% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{deviance.fitgen_glm}
\alias{deviance.fitgen_glm}
\title{Accesor for the deviance}
\usage{
\method{deviance}{fitgen_glm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_glm}}

\item{...}{Additional arguments}
}
\description{
Accesor for the deviance
}
