% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{deviance.fitgen_Lglm}
\alias{deviance.fitgen_Lglm}
\title{Accesor for deviances}
\usage{
\method{deviance}{fitgen_Lglm}(object, type = c("full", "common"), ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{The coefficients corresponds to the full model (with neighbours)
or to the common model (only with the common predictors)}

\item{...}{Additional arguments}
}
\description{
Accesor for deviances
}
