% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{dispersion.fitgen_Lglm}
\alias{dispersion.fitgen_Lglm}
\title{Accesor for dispersion}
\usage{
\method{dispersion}{fitgen_Lglm}(object, type = c("Bliss"), ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{Method to estimate the dispersions}

\item{...}{Additional arguments}
}
\description{
Accesors for dispersions
}
