% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{dispersion}
\alias{dispersion}
\title{dispersion}
\usage{
dispersion(object, type, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{Method to estimate the dispersions}

\item{...}{Additional arguments}
}
\description{
A function to create the S3 method dispersion
}
