% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitExtendedRowGLM}
\alias{FitExtendedRowGLM}
\title{FitExtendedRowGLM}
\usage{
FitExtendedRowGLM(dat, Cl, ng, model, commonX, offset, Phi = 0, retmu = FALSE)
}
\arguments{
\item{dat}{Numeric matrix of counts with n individuals/genes as rows and c studies/cases as columns}

\item{Cl}{Numeric matrix of affinity with n individuals/genes as rows and the integer indexes of the p closest individuals in its p columns}

\item{ng}{Number of the individual/gene for which the complete model matrix has to be built}

\item{model}{A string with the model to fit, either "poisson", "binlogit", "negbin" or "binlogitw"}

\item{commonX}{Partial model matrix with c rows and with the columns common to all individuals/genes. Usually generated as model.matrix(...)}

\item{offset}{Numeric vector with offsets for the GLM. Its length must be the number of colums of the counts matrix, i.e. c. If the data are normalized this vector should be passed filled with 0.}

\item{Phi}{Estimated value for Phi in the case of negative binomial model. It can't be 0.}

\item{retmu}{Boolean to indicate whether or not the vector of estimated means must be returned}
}
\value{
A list with the following keys:
\itemize{
\item model  - A string with the model used to adjust. A copy of the model parameter, useful to keep track of it when the result is included in a list 
\item beta   - A numeric vector with the value of the model parameters assuming the requested distribution
\item muest  - A numeric vector with the estimated means or NULL if parameter retmu was FALSE
\item X2     - The numeric value of the X2 statistic (see help of StGLM)
\item sums   - The numeric value of the sums statistic (see help of StGLM)
\item numnei - The number of neighbours that has been used to obtain a valid fit, either p or a lower value
\item Dev    - The numeric value of the deviance of the model
\item CovM   - Covariance matrix of the maximum likelihood estimator of the model coefficients
}
}
\description{
Fits a Poisson or Negative Binomial General Linear Model to the requested row of a data matrix using as predictors
those included in the commonX model matrix plus the columns of those rows in the data matrix which are neighbours of the current row.\cr
The concept of neighbourhood is encoded in a nxp matrix of closeness which contains int its i-th row\cr
the index of the p cases which are closest to the i-th case. The dimension of the matrix is used, so p is not explicitly passed as parameter.
}
\examples{
# To be done
}
