% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{new_fitgen_glm}
\alias{new_fitgen_glm}
\title{new_fitgen_glm}
\usage{
new_fitgen_glm(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional arguments}
}
\description{
A function to create the S3 object \code{fitgen_glm}
}
