% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitGenGetDebug}
\alias{FitGenGetDebug}
\title{FitGenGetDebug
Returns the current value of the debug flag (TRUE or FALSE).\cr
Used by the R functions of the package to know if they should emit messages or not.}
\usage{
FitGenGetDebug()
}
\value{
Boolean value with the current value of the internal DEB flag.
}
\description{
FitGenGetDebug
Returns the current value of the debug flag (TRUE or FALSE).\cr
Used by the R functions of the package to know if they should emit messages or not.
}
\examples{
if (FitGenGetDebug()) {
 cat("This is a message emitted because we are in debug mode.\n")
}
}
