% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitGenSetDebug}
\alias{FitGenSetDebug}
\title{FitGenSetDebug}
\usage{
FitGenSetDebug(deb, lldeb = FALSE)
}
\arguments{
\item{deb}{boolean, TRUE to generate normal debug messages and FALSE to turn them off.}

\item{lldeb}{boolean, TRUE to generate low level debug messages (very verbose) and FALSE to turn them off. Setting this parameter
to TRUE sets also deb to TRUE. Default: FALSE}
}
\value{
No return value, called for side effects (internal boolean flag changed)
}
\description{
Sets debugging in fitgen package to ON (with TRUE) or OFF (with FALSE) or even sets\cr
a lower level of debug to see inside the numerical algorithms.\cr
On package load the default status of both debugging levels is OFF.\cr
Setting any debug level to ON shows a message. Setting to OFF does not show anything (since debugging is OFF...)
}
\examples{
FitGenSetDebug(TRUE)
FitGenSetDebug(FALSE)
FitGenSetDebug(TRUE,TRUE)
FitGenSetDebug(TRUE,FALSE)
}
