% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitGLMInd}
\alias{FitGLMInd}
\title{FitGLMInd}
\usage{
FitGLMInd(
  dat,
  normalized,
  model,
  selected,
  commonX,
  Phi_estproc,
  Phi_external = as.numeric(c()),
  retmu = FALSE
)
}
\arguments{
\item{dat}{Numeric matrix of counts with n individuals/genes as rows and c studies/cases as columns}

\item{normalized}{Boolean value to indicate if data are passed normalized (TRUE) or not (FALSE). If not, offset  vector is calculated inside.}

\item{model}{A string with the model to fit, either "poisson","binlogit","negbin" or "binlogitw"}

\item{selected}{Boolean array of n positions indicating to which of the n rows the selected model should be fitted. Notice that calculation of
the offset vector, if requested by the normalized parameter, is done using all the individuals in the dat matrix, not only the selected ones.}

\item{commonX}{Partial model matrix with c rows and with the columns common to all individuals/genes. Usually generated as model.matrix(...)}

\item{Phi_estproc}{Procedure to estimate the Phi parameter in the case of negbinomial model. Possible values are "none","single","common","grouped","common_edger" or "tagwise_edger".}

\item{Phi_external}{Numeric vector with estimated Phi values if the estimation procedure is "common_edger" or "tagwise_edger". If a vector with correct length is passed, but estimation procedures has been set to any different to the former ones, the vector is used and the procedure is set to "common_edger". Default: empty numeric vector.}

\item{retmu}{Boolean to indicate wheter or not the vector of estimated means must be returned. Default: FALSE}
}
\value{
A list of lists, each one with the following keys:
\itemize{
\item model - A string with the model used to adjust. A copy of the model parameter, useful to keep track of it when the result is included in a list, or none if selected was FALSE for that row
\item beta0 - A numeric vector with the value of the model parameters
\item Dev0  - A numeric value with the deviance of the base model
\item CovM0 - A numeric matrix, the covariance matrix of the maximum likelihood estimator of the model coefficients for the base model
\item muest0 - A numeric vector with the estimated means, or NULL if parameter retmu was FALSE
\item iters0 - An integer with the number of iterations until convergence of the base model
\item Bliss - The numeric value of Phi obtained using the requested Bliss-Owen method (single, common or grouped) for fits to the negative binomial model or 0.0 otherwise
\item Wedderburn - The numeric value of the cuasidispersion parameter of the adjusted model according to the Wedderburn estimator (1979)
\item Farrington - The numeric value of the cuasidispersion parameter of the adjusted model according to the Farrington estimator (1996)
\item Fletcher   - The numeric value of the cuasidispersion parameter of the adjusted model according to the Fletcher estimator (2012)
}
}
\description{
Fits a General Linear Model to the requested rows of a data matrix by using as predictors those
included in the commonX model matrix. This is a particular case of FitGLMWithNei when no neighbours are to be considered.
}
\details{
It might be possible that the model cannot be adjusted. This manifests as the lack of convergence of the numerical
algorithms. In such a case the vector of coefficients beta0 is returned as a 0-length vector.
}
\examples{
# To be done
}
