% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitGLMWithNei}
\alias{FitGLMWithNei}
\title{FitGLMWithNei}
\usage{
FitGLMWithNei(
  dat,
  normalized,
  Cl,
  model,
  selected,
  commonX,
  Phi_estproc,
  Phi_external = as.numeric(c()),
  retmu = FALSE
)
}
\arguments{
\item{dat}{Numeric matrix of counts with n individuals/genes as rows and c studies/cases as columns}

\item{normalized}{Boolean value to indicate if data are passed normalized (TRUE) or not (FALSE). If not, offset  vector is calculated inside.}

\item{Cl}{Numeric matrix of affinity with n individuals/genes as rows and the integer indexes of the p closest individuals in its p columns}

\item{model}{A string with the model to fit, either "poisson","binlogit","negbin" or "binlogitw"}

\item{selected}{Boolean array of n positions indicating to which of the n rows the selected model should be fitted. Notice that calculation of
the offset vector, if requested by the normalized parameter, is done using all the individuals in the dat matrix, not only the selected ones.
Similarly, Cl must have the number of the closest neighbours, either if they are in the selected ones or not.}

\item{commonX}{Partial model matrix with c rows and with the columns common to all individuals/genes. Usually generated as model.matrix(...)}

\item{Phi_estproc}{Procedure to estimate the Phi parameter in the case of negbinomial model. Possible values are "none","single","common","grouped","common_edger" or "tagwise_edger".}

\item{Phi_external}{Numeric vector with estimated Phi values if the estimation procedure is "common_edger" or "tagwise_edger". If a vector with correct length is passed, but estimation procedures has been set to any different to the former ones, the vector is used and the procedure is set to "common_edger". Default: empty numeric vector.}

\item{retmu}{Boolean to indicat wheter or not the vectors of estimated means must be returned for both models. Default: FALSE}
}
\value{
A list of lists, each one with the following keys:
\itemize{
\item model - A string with the model used to adjust. A copy of the model parameter, useful to keep track of it when the result is included in a list, or none if selected was FALSE for that row
\item beta0 - A numeric vector with the value of the model parameters assuming a Poisson distribution and commonX as model matrix
\item Dev0  - A numeric value with the deviance of the base model
\item CovM0 - A numeric matrix, the covariance matrix of the maximum likelihood estimator of the model coefficients for the base model
\item muest0 - A numeric vector with the estimated means of the base model, or NULL if parameter retmu was FALSE
\item iters0 - An integer with the number of iterations until convergence of the base model
\item beta1 - A numeric vector with the value of the model parameters assuming a Poisson distribution and commonX plus as much as p added neighbours as model matrix.
\item Dev1  - A numeric value with the deviance of the extended model (commonX plus as much as p added neighbours)
\item CovM11 - A numeric matrix, the covariance matrix of the maximum likelihood estimator of the model coefficients for the extended model
\item muest1 - A numeric vector with the estimated means of the extended model, or NULL if parameter retmu was FALSE
\item iters1 - An integer with the number of iterations until convergence of the extended model
\item Bliss - The numeric value of Phi obtained using the requested Bliss-Owen method (single, common or grouped) for fits to the negative binomial model or 0.0 otherwise
\item Wedderburn - The numeric value of the cuasidispersion parameter of the adjusted model according to the Wedderburn estimator (1979)
\item Farrington - The numeric value of the cuasidispersion parameter of the adjusted model according to the Farrington estimator (1996)
\item Fletcher   - The numeric value of the cuasidispersion parameter of the adjusted model according to the Fletcher estimator (2012)
\item nnei  - An integer with the number of neigbours that still generate a valid adjusted model
}
}
\description{
Fits a General Linear Model to the requested rows of a data matrix by using as predictors those
included in the commonX model matrix and also those plus the columns of some rows in the data matrix which are neighbours of the current row.
}
\details{
The concept of neighbourhood between rows is encoded in a nxp matrix of closeness which contains in its i-th row
the index of the p cases which are closest to the i-th case. The function tries to adjust two models, one
with and one without the added neighbours and returns its parameters and measures of goodnes-of-fit.\cr
It might be possible that no model can be adjusted if too many predictors are incorporated in the model matrix,
or even without incorporating any additional predictors. This manifests as the lack of convergence of the numerical
algorithms. The parameter nnei returns the maximum number of neighbours (from p and downward) that allows a stable
adjustment. Value 0 would indicate that only the common model could be adjusted and in this case beta1 is NULL and
the Wedderburn and related parameters are those of the base model.
Value -1 is used to indicate that not even the common model fitted. In this case values of beta0 and beta1
are the empty vectors and overdispersion estimators are 0.
}
\examples{
# To be done
}
