% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitRowGLM}
\alias{FitRowGLM}
\title{FitRowGLM}
\usage{
FitRowGLM(dat, ng, model, commonX, offset, Phi = 0, retmu = FALSE)
}
\arguments{
\item{dat}{Numeric matrix of counts with n individuals/genes as rows and c studies/cases as columns}

\item{ng}{Number of the individual/gene for which the model must be adjusted}

\item{model}{A string with the model to fit, either "poisson","binlogit","negbin" or "binlogitw"}

\item{commonX}{Model matrix with c rows and with the columns common to all individuals/genes. Usually generated as model.matrix(...)}

\item{offset}{Numeric vector with offsets for the GLM. Its length must be the number of colums of the counts matrix, i.e. c. If the data are normalized this vector should be passed filled with 0.}

\item{Phi}{Estimated value for Phi in the case of negative binomial model. It can't be 0.}

\item{retmu}{Boolean to indicate whether or not the vector of estimated means must be returned. Default: FALSE}
}
\value{
A list with the following keys:
\itemize{
\item model - A string with the model used to adjust. A copy of the model parameter, useful to keep track of it when the result is included in a list 
\item beta  - A numeric vector with the value of the model parameters assuming the requested distribution
\item muest - A numeric vector with the estimated means or NULL if parameter retmu was FALSE
\item X2    - The numeric value of the X2 statistic (see help of StGLM)
\item sums  - The numeric value of the sums statistic (see help of StGLM)
\item Dev   - The numeric value of the deviance of the model
\item CovM  - Covariance matrix of the maximum likelihood estimator of the model coefficients
}
}
\description{
Fits a Poisson or General Linear Model to the requested row of a data matrix using as predictors
those included in the commonX model matrix
}
\details{
Error management: This function manages the errors in the same way as StGLM, i.e.: it returns an emtpy vector for beta,
the value 0 for X2, sums and Dev and the empty matrix for CovM if the model is numerically unstable. This MUST be checked
by the caller function and it is to allow it to call a different model if the current one cannot be fitted
}
\examples{
# To be done
}
