% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{fitted.fitgen_glm}
\alias{fitted.fitgen_glm}
\title{Accesor for the fitted values}
\usage{
\method{fitted}{fitgen_glm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_glm}}

\item{...}{Additional arguments}
}
\description{
Accesor for the fitted values
}
