% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GroupedCommonBlissOwen}
\alias{GroupedCommonBlissOwen}
\title{GroupedCommonBlissOwen}
\usage{
GroupedCommonBlissOwen(
  y,
  maxit = 100L,
  tol = 1e-09,
  tolphi = -1,
  posthres = 0,
  down_ndev = 3L,
  up_ndev = 2L,
  down_maxprop = 1,
  up_maxprop = 1,
  niter = 2L
)
}
\arguments{
\item{y}{A matrix of samples, size n x p}

\item{maxit}{Maximum number of allowed iterations for each Phi estimation. Default: 100}

\item{tol}{Tolerance to exit if change in each Phi estimation is below its value. Default: 1e-9}

\item{tolphi}{Maximum absolute difference in Phi value to be part of the estimation around a given sample. Default: -1.0, which means 'Estimate from the sample'}

\item{posthres}{Threshold to consider a sample valid if its v-value is strictly above it. Default: 0.0}

\item{down_ndev}{Integer with the number of stdev to limit interval of grouping by the lower side. Default: 3}

\item{up_ndev}{Integer with the number of stdev to limit interval of grouping by the upper side. Default: 2}

\item{down_maxprop}{Maximum proportion of the total samples that, being in the interval [mean-down_ndev*stdev,mean], must be considered. Default: 1.0 (consider all the samples that fall in the interval)}

\item{up_maxprop}{Maximum proportion of the total samples that, being in the interval [mean,mean+up_ndev*stdev], must be considered. Default: 1.0 (consider all the samples that fall in the interval)}

\item{niter}{Integer with the number of iteration smoothing steps. Default: 2}
}
\value{
A list L of n lists, each one (i.e.: L[[row]]) with the following keys:
\itemize{
          \item Phi    - A numeric value with the estimated Phi value for L[[row]]
          \item valid  - A boolean indicating if row has been kept or discarded according to the v-threshold criterion
          \item nsel   - An integer indicating how many samples were selected to be averaged with the sample at this row
          \item tolphi - A numeric value with the estimated tolerance for sample grouping for this row
}
If L[[i]]$valid is false the values of L[[i]]$Phi$ and of L[[i]]$tolphi will be NaN and the value of L[[i]]$nsel will be 0.
}
\description{
Gets a matrix of n samples of p observations each given as a nxp matrix (one sample per row) and, assuming that all 
of them come from several negative binomial ditributions, each with a different overdispersion Phi value
(var_i = mu_i+Phi_i*mu_i^2), uses the Bliss-Owen method to get estimations for the different Phi values when they
can be estimated (which is determined by the fact that its v-statistic is strictly above a threshold, 0 by default).
}
\details{
The estimation of each Phi proceeds by estimating first the value for each row with the simple BlissOwen method and
then, grouping for each sample all the others whose absolute value of difference of Phi is less than the tolerance and
using them to estimate the value through the same method used in the CommonBlissOwen function programmed in this package.\cr
The default value for tolphi, which is -1.0, does not mean -1 (which would make no sense) but a mark to indicate that\cr
it must be estimated using from formula (5) of BlissOwen[58], which depends on the sample. If you set any other value\cr
for tolphi that value will be used for all rows instead of individually estimating the tolerance for each sample.
}
\examples{
n = 50
size = seq(.2,2,length.out=5)
mu = 30
y = t(sapply(size,function(i) rnbinom(n = n, size = i, mu= mu)))
VPhi=GroupedCommonBlissOwen(y)
}
