% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{new_fitgen_Lglm}
\alias{new_fitgen_Lglm}
\title{Class fitgen_Lglm}
\usage{
new_fitgen_Lglm(object, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{...}{Additional arguments}
}
\description{
Class fitgen_Lglm
}
