% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{QRdec}
\alias{QRdec}
\title{QRdec}
\usage{
QRdec(A, tol = 1e-09)
}
\arguments{
\item{A}{A (m x n) numeric matrix to be decomposed with m>=n}

\item{tol}{Numeric tolerance for the norm of the Housholder vector. Possitive number. Default: 1e-9}
}
\value{
ret   A list with two elements
              ret["Q"] The orthogonal matrix Q
              ret["R"] The upper-triangular matrix R
}
\description{
Makes the QRdecomposition of a matrix A as A=Q*R using the Householder method.\cr
}
\details{
WARNING: if you are trying to solve a linear system of the form Ax=B, DO NOT use this function\cr
Use instead SolveLinearSystem. Both operate with the same algorithm and are equally efficient in
the decomposition but this one explicitly constructs Q and R which uses more memory.\cr

WARNING: the passed matrix IS CHANGED inside this function. By efficiency, we have opted for NOT copying
the argument in a new, temporary matrix inside the function. If we want to keep your original A make a 
deep copy with\cr
                  Acopy<-matrix(A,nrow=m,ncol=n)\cr
before calling this function.\cr
}
\examples{
m=10
n=8
A=matrix(runif(m*n),nrow=m,ncol=n)
Acop=matrix(A,nrow=m,ncol=n)
ret <- QRdec(A)
Arec <- ret$Q  \%*\% ret$R
error <- max(abs(Arec-Acop))
}
