% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{quasidispersion.fitgen_Lglm}
\alias{quasidispersion.fitgen_Lglm}
\title{Accesor for quasidispersions}
\usage{
\method{quasidispersion}{fitgen_Lglm}(object, type = c("Wedderburn", "Farrington", "Fletcher"), ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{Method of estimation of the quasidispersions}

\item{...}{Additional arguments}
}
\description{
Accesor for quasidispersions
}
