% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgen_S3classes.R
\name{quasidispersion}
\alias{quasidispersion}
\title{quasidispersion}
\usage{
quasidispersion(object, type, ...)
}
\arguments{
\item{object}{Object of class \code{fitgen_Lglm}}

\item{type}{Method of estimation of  the quasidispersions}

\item{...}{Additional arguments}
}
\description{
A function to create the S3 method quasidispersion
}
