% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomMatrix.R
\name{rmatNB2}
\alias{rmatNB2}
\title{rmatNB2}
\usage{
rmatNB2(N, x, common_beta, neighbours_beta = NULL, phi, seed = 1234)
}
\arguments{
\item{N}{Number of features (genes)}

\item{x}{Model matrix}

\item{common_beta}{Coefficients correspondings to \code{x} i.e.
the common predictors for all rows}

\item{neighbours_beta}{Coefficients of neighbours}

\item{phi}{The dispersion parameters (Nx1)}

\item{seed}{The seed for the random generator}
}
\description{
rmatNB2
}
