% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SolveLinearSystem}
\alias{SolveLinearSystem}
\title{SolveLinearSystem}
\usage{
SolveLinearSystem(A, b, tol = 1e-09)
}
\arguments{
\item{A}{Numeric matrix (system matrix)}

\item{b}{Numeric vector (independent term)}

\item{tol}{Numeric tolerance for the norm of the Housholder vector. Possitive number. Default: 1e-9}
}
\value{
x      A numeric vector of length n with the solution\cr
}
\description{
Solves the linear system Ax=b where A is a numeric matrix of size m x n (with n>1), b is a column vector of length m
and x (the solution) is a column vector of length n.
 
The solution is obtained in the minimal square error sense (i.e.: sum_i |Ax_i-b|^2 is a minimum) using
non-explicit QR factorization of A. Therefore, the number of equations must be bigger or equal than the number
of unknows so m must be >= n. A is supposed to be well conditioned (i.e.: rank n). Numerical problems may arise
for badly conditioned A (see Error management)\cr
}
\details{
WARNING: matrix A IS CHANGED inside this function. By efficiency, we have opted for NOT copying the argument in
a new, temporary matrix inside the function. If we want to keep your original A make a deep copy with\cr
                  Acopy<-matrix(A,nrow=m,ncol=n)\cr
before calling this function.\cr

After the call A will contain the upper-triangular R matrix in its upper-diagonal portion and the second-to-final
components of each Housholder vector v_i in the lower-diagonal part of column i. The first component is not stored
(no place to keep it) but it can be recovered since the Housholder vectors are of modulus 1. The Householder
matrix H_i can be constructed as\cr
                 H_i = I(m)-2*v_i'*v_i\cr
being I(m) the identity matrix of size n x n and v_i' the transposed of v_i.\cr
Finally, if the Q matrix is needed, it is\cr
                 Q=H(n-1) H(n-2) ... H_0 (in that order).\cr
Indeed, system is solved writing A*x*b as\cr
                 Q*R*x=b ==> R*x = Q'*b = H(n-1)*H(n-2)* ... *H_0 b\cr
(notice that inv(Q)=Q' and that H'(i)=H(i)\cr
Multiplications are done accumulatively as\cr
                 b(1)=H(1)*b, b(2)=H(2)*b(1), b(n-1)=H(n-1)*b(n-2), ...\cr
and the final triangular system\cr
                 R*x=b(n-1)\cr
is solved by back-susbtitution.

Error management:\cr
This function calls Rcpp::stop to indicate argument passing errors but it returns an empty vector to indicate lack
of solution because of numerical unstability, i.e.: the norm of a Housholder vector is below the tolerance. This
is to allow other functions try with different models if one of them cannot be adjusted. This means that the caller
function SHOULD check the length of the returned solution vector.
}
\examples{
m=9
n=6
A=matrix(nrow=m,ncol=n)
for (i in c(1:n))
{
 A[,i]=c(rep(0,i-1),c(1:(9-i+1)))
}
b=as.vector(c(m:1))
x=SolveLinearSystem(A,b)

}
