% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SolveLinearSystemWithInverse}
\alias{SolveLinearSystemWithInverse}
\title{SolveLinearSystemWithInverse}
\usage{
SolveLinearSystemWithInverse(A, b, tol = 1e-09)
}
\arguments{
\item{A}{Numeric matrix (system matrix)}

\item{b}{Numeric vector (independent term)}

\item{tol}{Numeric tolerance for the norm of the Housholder vector. Possitive number. Default: 1e-9}
}
\value{
A list with the following keys:
\itemize{
\item x      A numeric vector of length n with the solution\cr
\item Inv    A numeric matrix with the inverse of A\cr
}
}
\description{
Solves the linear system Ax=b where A is a square numeric matrix of size n x n (with n>1), b is a column vector of length n
and x (the solution) is a column vector of length n. It also returns the inverse of matrix A which, if used for estimation
of a linear model, is the variance/covariance matrix. Inversion is done using the QR decomposition of A used to solve the
linear system, so it should be faster than using simply SolveLinearSystem and invert the A matrix later.
 
The solution is obtained using QR factorization of A. A is supposed to be well conditioned (i.e.: rank n). Numerical problems
may arise for badly conditioned A (see Error management)\cr
}
\details{
WARNING: matrix A IS CHANGED inside this function. By efficiency, we have opted for NOT copying the argument in
a new, temporary matrix inside the function. If you want to keep your original A make a deep copy with\cr
                  Acopy<-matrix(A,nrow=m,ncol=n)\cr
before calling this function.\cr

After the call A will contain the upper-triangular R matrix in its upper-diagonal portion and the second-to-final
components of each Housholder vector v_i in the lower-diagonal part of column i. The first component is not stored
(no place to keep it) but it can be recovered since the Housholder vectors are of modulus 1. The Householder
matrix H_i can be constructed as\cr
                 H_i = I(m)-2*v_i'*v_i\cr
being I(m) the identity matrix of size n x n and v_i' the transposed of v_i.\cr
Finally, if the Q matrix is needed, it is\cr
                 Q=H(n-1) H(n-2) ... H_0 (in that order).\cr
Indeed, system is solved writing A*x*b as\cr
                 Q*R*x=b ==> R*x = Q'*b = H(n-1)*H(n-2)* ... *H_0 b\cr
(notice that inv(Q)=Q' and that H'(i)=H(i)\cr
Multiplications are done accumulatively as\cr
                 b(1)=H(1)*b, b(2)=H(2)*b(1), b(n-1)=H(n-1)*b(n-2), ...\cr
and the final triangular system\cr
                 R*x=b(n-1)\cr
is solved by back-substitution.

Inverse of A is done using the QR-decomposition calculated to solve the system. Indeed, if\cr
A=Q*R then\cr
A^-1=R^-1*Q^-1.\cr
Being Q orthogonal, Q^-1 = Q^T, and for the upper-triangular matrix R, it can be written as\cr
R=D(I+N) where D is a diagonal matrix and N is nilpotent (i.e.: N^n=0). In such a case,\cr
R^-1=(I+N)^-1*D^-1.\cr
where, since D is diagonal, calculating D^-1 is trivial and a matrix lemma stablishes that, for nilpotent matrices N,\cr
(I+N)^-1=I+sum_p=1..(n-1) N^p\cr

Error management:\cr
This function calls Rcpp::stop to indicate argument passing errors but it returns an empty vector to indicate lack
of solution because of numerical unstability, i.e.: the norm of a Housholder vector is below the tolerance. This
is to allow other functions try with different models if one of them cannot be adjusted. This means that the caller
function SHOULD check the length of the returned solution vector.
}
