% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StGLM}
\alias{StGLM}
\title{StGLM}
\usage{
StGLM(y, X, model, offset, warn_me, Phi = 0, retmu = FALSE)
}
\arguments{
\item{y}{A numeric vector of observations, length m}

\item{X}{A numeric matrix of predictors, size m x n}

\item{model}{A string with the model to use. Possible values: 'Poisson', 'BinLogit', 'NegBin' or 'BinLogitW'\cr
for Possion, Binomial with logistic link, Negbinomial, and Grouped Binomial with logistic link, respectively.}

\item{offset}{A numeric vector of offsets, length m}

\item{warn_me}{A logical value indicating if a warning must be raised in the case of numerical unstability}

\item{Phi}{A real value, the overdispersion parameter. Used only for Negbinomial. Default: 0.0}

\item{retmu}{A boolean value to indicate whether or not the vector of estimated means must be returned. Default: FALSE}
}
\value{
A list with the following keys:
\itemize{
\item beta  		A numeric vector with the estimated model coefficients, or empty vector if they cannot be estimated
\item muest         A numeric vector with the estimated means, or NULL if parameter retmu was FALSE
\item X2            The statistic X^2, defined as sum((y_i-muest_i)^2/Vest_i). 0 if the model cannot be estimated
\item sums          The statistic sums, defined as sum((y_i-muest_i)*Vest'_i/Vest_i). 0 if the model cannot be estimated
\item Dev			The deviance of the model. 0 if the model cannot be estimated
\item CovM          Variance-covariance matrix of the maximum likelihood estimator of the model coefficients. Empty matrix if the model cannot be estimated
}
}
\description{
Gets a vector of observations, a type of link function, the phi parameter, the offset
vector and the Phi-star parameter and returns the vector of estimators beta, the value of X^2 and the final
values of the deviance and of the covariance matrix of the maximum likelihood estimator of the model coefficients
}
\details{
Error management:\cr
This function calls Rcpp::stop only to signal error in pass of parameters. Errors due to numerical unstability
are not raised (at most, a warning is called, see parameter warn_me) but function returns an empty vector as beta value,
0 for X2, sums and Dev and the empty matrix for CovM. This MUST be checked by the caller function and it is to allow it
to test a different, simpler model if the current one cannot be fitted.
}
\examples{
# To be done
}
